Made the first succesfull execution loop
I made a while loops that iterates trough the minishell struct and uses a system of two pipes. One to redirect the output of the current when there is a pipe to the left and one that saves the output of the current command when there is a pipe to the left. It seems to be working suprizingly well and now I think I will start working on a sistem that will update the env table in parent process using a signals and pipes combination.